+++ markdown-multiplatform-html/src/commonMain/kotlin/com/iffly/compose/markdown/multiplatform/html/handlers/DefaultHtmlInlineTagHandlers.kt
@@ -74,9 +74,7 @@ internal class LinkTagHandler : HtmlInlineTagHandler {
     ) {
         val url = HREF_REGEX.find(rawTag)?.groupValues?.get(1) ?: ""
         val linkInteractionListener =
-            context.actionHandler?.let {
-                MarkdownLinkInteractionListener(actionHandler = it, node = context.node)
-            }
+            MarkdownLinkInteractionListener(actionHandler = context.actionHandler, node = context.node)
         val linkAnnotation =
             LinkAnnotation.Url(
                 url = url,