+++ markdown-multiplatform/src/commonMain/kotlin/com/iffly/compose/markdown/multiplatform/config/LocalActionHandlerProvider.kt
@@ -4,12 +4,13 @@ package com.iffly.compose.markdown.multiplatform.config
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.staticCompositionLocalOf
-import com.iffly.compose.markdown.multiplatform.ActionHandler
+import com.iffly.compose.markdown.multiplatform.ActionHandlerState
 
 internal val LocalActionHandlerProvider =
-    staticCompositionLocalOf<ActionHandler?> {
-        null
+    staticCompositionLocalOf<ActionHandlerState> {
+        mutableStateOf(null)
     }
 
 /**
@@ -19,4 +20,4 @@ internal val LocalActionHandlerProvider =
  */
 @Composable
 @ReadOnlyComposable
-fun currentActionHandler(): ActionHandler? = LocalActionHandlerProvider.current
+fun currentActionHandler(): ActionHandlerState = LocalActionHandlerProvider.current